izmjena rezervacija

+++ isa/src/main/java/rs/ftn/isa/model/Usluga.java
@@ -120,7 +120,7 @@ public class Usluga {
 	}
 
 
-	public int getPopust() {
+	public double getPopust() {
 		return popust;
 	}
